Type:fix
Description:bug#19855 After clicking the "Add Server" and "Edit Server" buttons in the remote management panel, it will take a few seconds to display.

RootCause:Read the configuration file to get the group name during initialization.

Solution:Do not read the configuration to obtain data when the pop-up window is initialized.