Type:fix
Description:BUG#22626 Enter Chinese content in the search box, search
content matches incorrectly

RootCause:Null

Solution:Save the relevant value of each search, and then add the number
of Chinese characters to calculate the selection area. When searching up
or down again, the saved value will be used as the starting position for
the next search.