Type:fix
Description:bug#28221 When there is no server in the group, it is recommended to jump directly to the remote management list.

RootCause:When there is no server in the group, there is no jump to the remote management list.

Solution:
Pànduàn fēnzǔ shìfǒu yǒu fúwùqì, méiyǒu jiù fǎnhuí
您是不是要找： 判断分区是否有服务器,没有就返回
16/5000
Determine if the packet has a server, return if not